Refactor utility functions inside checks result module
