+++ assets/js/components/ChecksResults/index.js
@@ -2,7 +2,16 @@ import ChecksResults from './ChecksResults';
 import ResultsContainer from './ResultsContainer';
 import HostResultsWrapper from './HostResultsWrapper';
 import CheckResult from './CheckResult';
+import { getHosts, getChecks, getHealth, getCheckResults } from './checksUtils';
 
-export { ResultsContainer, HostResultsWrapper, CheckResult };
+export {
+  ResultsContainer,
+  HostResultsWrapper,
+  CheckResult,
+  getHosts,
+  getChecks,
+  getHealth,
+  getCheckResults,
+};
 
 export default ChecksResults;
